[scene] Move Sampler out of Texture.
